requirements.txt
aiogram==3.0.0
python-decouple==3.8
